+++ src/main/java/guideme/document/block/AlignItems.java
@@ -1,7 +1,21 @@
 package guideme.document.block;
 
-public enum AlignItems {
+import java.util.Locale;
+import net.minecraft.util.StringRepresentable;
+
+public enum AlignItems implements StringRepresentable {
     CENTER,
     START,
-    END
+    END;
+
+    private final String serializedName;
+
+    AlignItems() {
+        this.serializedName = name().toLowerCase(Locale.ROOT);
+    }
+
+    @Override
+    public String getSerializedName() {
+        return serializedName;
+    }
 }